minor prettying

+++ app/models/group.rb
@@ -6,43 +6,32 @@ class Group < ApplicationRecord
 
   validates_presence_of :label
 
-
   #Alle "Fremd-Shares" löschen, wenn Klasse archiviert wird
-  after_save do |g|
-    if g.archive
-      GroupShare.where(group: g, owner: false).destroy_all
-    end
-  end
+  after_save { |g| GroupShare.where(group: g, owner: false).destroy_all if g.archive }
 
   #Alle Einträge aus Share löschen, bevor Klasse gelöscht wird
-  before_destroy do |g|
-    GroupShare.where(group: g).destroy_all
-  end
+  before_destroy { |g| GroupShare.where(group: g).destroy_all }
 
   #Liefere Objekt für Frontend, das Group und GroupShare vereint
   def as_hash(user)
-    c = group_shares.where(user: user).first  #Kombination aus User & Group muss eindeutig sein!
+    c = group_shares.where(user: user).first #Kombination aus User & Group muss eindeutig sein!
     data = {
-        id: id,
-        label: label,
-        size: students.size,
-        students: students,
-        demo: demo,
-        auth_token: auth_token,
-        read_only: c.read_only,
-        archive: archive,
-        updated_at: updated_at,
-        key: c.key,
-        owner: c.owner,
-        share_id: c.id
+      id: id,
+      label: label,
+      size: students.size,
+      students: students,
+      demo: demo,
+      auth_token: auth_token,
+      read_only: c.read_only,
+      archive: archive,
+      updated_at: updated_at,
+      key: c.key,
+      owner: c.owner,
+      share_id: c.id
     }
     data['belongs_to'] = c.group.owner.email
     data['shares'] = []
-    group_shares.all.each do |c|
-      if c.user != user
-        data['shares'] += [c.as_hash]
-      end
-    end
+    group_shares.all.each { |c| data['shares'] += [c.as_hash] if c.user != user }
     return data
   end
 
@@ -54,5 +43,4 @@ class Group < ApplicationRecord
   def read_only(user)
     group_shares.where(user: user).pluck(:read_only).first
   end
-
 end
